requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
seaborn==0.13.2
matplotlib==3.8.3
pymongo==4.6.1
python-dotenv==1.0.1
scikit-learn==1.3.2
dask==2024.2.1
pyarrow==15.0.0
geopandas==0.14.3
kaleido==0.2.1
psutil==5.9.8